Factor out build app core logic to lib
