+++ server/Cargo.toml
@@ -25,3 +25,7 @@ tower-service = "0.3"
 tracing = { version = "0.1", features = ["log"] }
 tracing-subscriber = { version = "0.3", default-features = false, features = ["fmt", "env-filter"] }
 ulid = { version = "1.0.0", features = ["serde"] }
+
+[[bin]]
+name = "lambda"
+path = "./src/lambda.rs"
