+++ server/src/toggle.rs
@@ -1,5 +1,5 @@
-use super::{Backend, Local};
-use crate::to_dynamo_timestamp;
+use crate::utils;
+use crate::{Backend, Local};
 use aws_sdk_dynamodb::{
     error::SdkError,
     operation::update_item::{UpdateItemError, UpdateItemOutput},
@@ -52,7 +52,10 @@ impl Backend {
                         if let Some(time) = time {
                             q.update_expression("SET #field = :set")
                                 .expression_attribute_names("#field", "answered")
-                                .expression_attribute_values(":set", to_dynamo_timestamp(time))
+                                .expression_attribute_values(
+                                    ":set",
+                                    utils::to_dynamo_timestamp(time),
+                                )
                         } else {
                             q.update_expression("REMOVE #field")
                                 .expression_attribute_names("#field", "answered")
@@ -72,7 +75,7 @@ impl Backend {
                     ToggleRequest::Hidden(set) => q.insert("hidden", AttributeValue::Bool(set)),
                     ToggleRequest::Answered(time) => {
                         if let Some(time) = time {
-                            q.insert("answered", to_dynamo_timestamp(time))
+                            q.insert("answered", utils::to_dynamo_timestamp(time))
                         } else {
                             q.remove("answered")
                         }
@@ -90,7 +93,7 @@ pub(super) async fn toggle(
     State(dynamo): State<Backend>,
     body: String,
 ) -> Result<Json<serde_json::Value>, StatusCode> {
-    super::check_secret(&dynamo, &eid, &secret).await?;
+    utils::check_secret(&dynamo, &eid, &secret).await?;
 
     let req = match (&*body, property) {
         ("on", Property::Hidden) => ToggleRequest::Hidden(true),
